+++ info/ChangeLog.md
@@ -1,3 +1,10 @@
 # Change Log
 
 Where changes are placed
+
+## 0.1
+
+* Added actor resource
+* Added player
+* Added pointer
+* Added stats manager